expert_system/pyproject.toml
[tool.poetry]
name = "expert-system"
version = "0.1.0"
description = ""
authors = ["Александр Лыо <al.lyo@yandex.ru>"]
readme = "README.md"
packages = [{include = "expert_system"}]

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.1.1"
django = "^4.2.2"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
django-tinymce = "^3.6.1"
djangorestframework = "^3.14.0"
isort = "^5.12.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
